Deliverable_2/FastAPI/setting.py
import os

from dotenv import load_dotenv

fastapi_env_path = 'Deliverable_2/FastAPI/.env'
load_dotenv(fastapi_env_path)

POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
POSTGRES_USER = os.getenv("POSTGRES_USER")
POSTGRES_DB = os.getenv("POSTGRES_DB")
POSTGRES_PORT = os.getenv("POSTGRES_PORT")
POSTGRES_HOST = os.getenv("POSTGRES_HOST")
