Deliverable_2/FastAPI/requirements.txt
fastapi==0.110.1
pydantic==2.6.4
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
